1、分页查询nextSize改为int类型

+++ wolf-framework/src/main/java/com/wolf/framework/service/request/ListRequest.java
@@ -8,5 +8,5 @@ public interface ListRequest extends ObjectRequest {
 
     public long getNextIndex();
     
-    public long getNextSize();
+    public int getNextSize();
 }
